chore(admin-skills): Admin skills page show if skill linked to request (#72)
* feat(admin-skills): add getRequestSkill method to get data from request_skills table

* feat(admin-skills): add request-skills endpoint

* feat(admin-skills): delete 'request-skills' endpoint

* feat(admin-skills): delete 'getRequestSkill' method

* feat(admin-skills): add 'requestSkills' relationship method

* feat(admin-skills): refactor 'all' method 'else return' statement to include 'request skills count'

* feat(admin-skills): add 'user_skills_count' to '/skills GET' endpoint response

* feat(admin-skills): add can't delete skill that's in use edge case to 'remove' method

* feat(admin-skills): order skills by name on 'admin-skills' page

+++ app/Http/Controllers/SkillController.php
@@ -40,7 +40,7 @@ class SkillController extends Controller
 
             return $this->respond(Response::HTTP_OK, ["data" => $matched_skill]);
         } else {
-            return $this->respond(Response::HTTP_OK, ["data" => $m::orderBy('name')->get()]);
+            return $this->respond(Response::HTTP_OK, ["data" => $m::withCount('requestSkills', 'userSkills')->orderBy('name')->get()]);
         }
     }
 
@@ -134,6 +134,12 @@ class SkillController extends Controller
     public function remove(Request $request, $id)
     {
         try {
+            if (UserSkill::where('skill_id', $id)->exists()
+                || \App\RequestSkill::where('skill_id', $id)->exists()
+            ){
+                return $this->respond(Response::HTTP_FORBIDDEN, ["data" => "Skill is currently in use"]);
+            }
+            
             $skill = Skill::findOrFail($id);
             $skill->delete();
